#208: Refactored the full-stack app to use the Vite plugin
