+++ examples/3-apps/1-contact-list/conf/dev.json
@@ -1,0 +1,10 @@
+{
+    "url": "http://127.0.0.1:3000",
+    "standalone":
+    {
+        "source": "./dist",
+        "cache": "./cache",
+        "index": "index.html",
+        "assets": [ "index.html", "main.js", "App.js", "vite.svg", "assets/**/*" ]
+    }
+}
