+++ examples/3-apps/1-contact-list/conf/gateway.json
