+++ examples/3-apps/1-contact-list/conf/repo.json
@@ -1,0 +1,10 @@
+{
+    "url": "http://127.0.0.1:2999",
+    "repository":
+    {
+        "source": "./dist",
+        "cache": "./cache",
+        "index": "index.html",
+        "assets": [ "index.html", "main.js", "App.js", "vite.svg", "assets/**/*" ]
+    }
+}
