+++ examples/3-apps/1-contact-list/docker-compose.yml
