+++ examples/3-apps/1-contact-list/src/App.tsx
@@ -1,0 +1,15 @@
+
+import Header from './components/Header';
+import ContactPage from './pages/ContactPage';
+
+function App()
+{
+  return (
+    <main className='container'>
+      <Header />
+      <ContactPage />
+    </main>
+  )
+}
+
+export default App;
