+++ examples/3-apps/1-contact-list/src/shared/common/DatabaseError.ts
