+++ examples/3-apps/1-contact-list/src/shared/contact/Contact.ts
