+++ examples/3-apps/1-contact-list/src/shared/contact/deleteContact.ts
@@ -1,0 +1,13 @@
+
+import getCollection from '../common/getCollection';
+import createId from '../common/createId';
+
+import Contact from './Contact';
+
+export default async function deleteContact(contact: Contact): Promise<void>
+{
+    const collection = await getCollection('contacts');
+    const mongoId = await createId(contact.id);
+
+    await collection.deleteOne({ _id: mongoId });
+}
