+++ examples/3-apps/1-contact-list/vite.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import jitar from '@jitar/plugin-vite';
+
+export default defineConfig({
+  plugins: [
+    react(),
+    jitar('src', 'shared', 'http://localhost:3000')
+  ]
+})
