--- examples/3-apps/1-full-stack/conf/dev.json
@@ -1,10 +1,0 @@
-{
-    "url": "http://127.0.0.1:3000",
-    "standalone":
-    {
-        "source": "./dist",
-        "cache": "./cache",
-        "index": "index.html",
-        "assets": [ "*.html", "*.js" ]
-    }
-}
