--- examples/3-apps/1-full-stack/conf/repo.json
@@ -1,10 +1,0 @@
-{
-    "url": "http://127.0.0.1:2999",
-    "repository":
-    {
-        "source": "./dist",
-        "cache": "./cache",
-        "index": "index.html",
-        "assets": [ "*.html", "*.js" ]
-    }
-}
