--- examples/3-apps/1-full-stack/global.d.ts
@@ -1,7 +1,0 @@
-import react from "react";
-import react_dom from "react-dom/client";
-
-declare global {
-    type React = typeof react
-    type ReactDOM = typeof react_dom
-}
