--- examples/3-apps/1-full-stack/package.json
@@ -1,28 +1,0 @@
-{
-    "name": "jitar-full-stack-example",
-    "type": "module",
-    "private": true,
-    "scripts": {
-        "clean": "rimraf dist",
-        "compile": "tsc",
-        "copy": "cpx -u 'src/index.html' dist",
-        "build": "npm-run-all clean copy compile",
-        "update": "npm-run-all copy compile",
-        "dev": "node --experimental-network-imports dist/server.js --config=./conf/dev.json",
-        "repo": "node --experimental-network-imports dist/server.js --config=./conf/repo.json",
-        "gateway": "node --experimental-network-imports dist/server.js --config=./conf/gateway.json",
-        "node": "node --experimental-network-imports dist/server.js --config=./conf/node.json",
-        "proxy": "node --experimental-network-imports dist/server.js --config=./conf/proxy.json"
-    },
-    "dependencies": {
-        "mongodb": "^5.1.0",
-        "jitar": "^0.4.0"
-    },
-    "devDependencies": {
-        "@types/react": "^18.0.28",
-        "@types/react-dom": "^18.0.11",
-        "cpx2": "^4.2.0",
-        "npm-run-all": "^4.1.5",
-        "rimraf": "^4.1.2"
-    }
-}
