--- examples/3-apps/1-full-stack/segments/server.segment.json
@@ -1,11 +1,0 @@
-{
-    "./app/common/database/createId": { "default": { "access": "private" } },
-    "./app/common/database/getCollection": { "default": { "access": "private" } },
-    "./app/common/database/getDatabase": { "default": { "access": "private" } },
-
-    "./app/contact/actions/createContact": { "default": { "access": "public" } },
-    "./app/contact/actions/deleteContact": { "default": { "access": "public" } },
-    "./app/contact/actions/getContact": { "default": { "access": "public" } },
-    "./app/contact/actions/getContacts": { "default": { "access": "public" } },
-    "./app/contact/actions/updateContact": { "default": { "access": "public" } }
-}
