--- examples/3-apps/1-full-stack/src/app/App.tsx
@@ -1,13 +1,0 @@
-
-import Header from './common/views/Header';
-import ContactPage from './contact/views/ContactPage';
-
-export function App(props: any)
-{
-    return (
-        <main className='container'>
-            <Header />
-            <ContactPage />
-        </main>
-      );
-}
