--- examples/3-apps/1-full-stack/src/client.tsx
@@ -1,14 +1,0 @@
-
-import { startClient } from 'jitar';
-
-const client = await startClient();
-
-// We need to import the App component with the Jitar client.
-// This enables the segmentation of the application.
-const { App } = await client.import('./app/App') as any;
-
-const rootElement = document.getElementById('root') as HTMLElement;
-
-// @ts-ignore (avoids the need for importing ReactDOM)
-const root = ReactDOM.createRoot(rootElement);
-root.render(<App />);
