--- examples/3-apps/1-full-stack/src/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-    <head>
-        <title>Full Stack example | Jitar</title>
-        <meta charset="utf-8">
-        <script crossorigin src="https://unpkg.com/react@18/umd/react.production.min.js"></script>
-        <script crossorigin src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js"></script>
-        <link crossorigin rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css">
-        <link crossorigin rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.9.1/font/bootstrap-icons.css">
-        <script type="importmap">
-            {
-                "imports":
-                {
-                    "jitar": "/jitar/client.js",
-                    "@jitar/runtime": "/jitar-runtime/lib.js",
-                    "@jitar/serialization": "/jitar-serialization/lib.js",
-                    "@jitar/reflection": "/jitar-reflection/lib.js"
-                }
-            }
-        </script>
-    </head>
-    <body>
-        <div id="root"></div>
-        <script src="./client.js" type="module"></script>
-    </body>
-</html>
