--- examples/3-apps/1-full-stack/tsconfig.json
@@ -1,16 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2022",
-    "jsx": "react",
-    "module": "es2022",
-    "rootDir": "./src/",
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "outDir": "./dist/",
-    "removeComments": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "skipLibCheck": true
-  },
-  "exclude": ["cache", "dist", "node_modules"]
-}
