+++ examples/README.md
@@ -14,14 +14,15 @@ In order to run the examples you need to have the following installed:
 * [Node.js](https://nodejs.org/en/) version 18.7.0 or higher
 * [TypeScript](https://www.typescriptlang.org/) version 4.4.2 or higher
 
-All examples come with a ``requests.http`` file containing example requests to call the procedures. We use
+Most examples do not have an UI to keep them simple and focussed. For running these examples the PRC API will be used.
+A ``requests.http`` file is provided containing example requests to call the procedures. We use
 [Visual Studio Code](https://code.visualstudio.com/) with the
 [REST Client](https://marketplace.visualstudio.com/items?itemName=humao.rest-client) extension to run these requests.
 But you can use any http client like [Postman](https://www.postman.com/).
 
 ## Basic examples
 
-The following examples are the most basic examples to get started with Jitar. They demonstrate the follwing basic features.
+The following examples are the most basic examples to get started with Jitar. They demonstrate the following basic features.
 
 1. [Hello Word](1-basic/1-hello-world/README.md) - Demonstrates how to create a procedure and call it using the RPC API.
 1. [Segmentation](1-basic/2-segmentation/README.md) - Demonstrates how to split an application into multiple segments.
@@ -34,7 +35,7 @@ The following examples are the most basic examples to get started with Jitar. Th
 
 ## Advanced examples
 
-The following examples are more advanced examples to get started with Jitar. They demonstrate the follwing advanced features.
+The following examples are more advanced examples to get started with Jitar. They demonstrate the following advanced features.
 
 1. [Start hooks](2-advanced/1-start-hooks/README.md) - Demonstrates how to start a Jitar server and client.
 1. [Run procedure](2-advanced/2-run-procedure/README.md) - Demonstrates how to run a procedure dynamically.
@@ -46,5 +47,5 @@ The following examples are more advanced examples to get started with Jitar. The
 The following examples combine the basic and advanced features. They demonstrate how to build real-world
 applications using Jitar.
 
-1. [Full stack](3-apps/1-full-stack/README.md) - Demonstrates how to build a full stack application using React, MongoDB and Jitar.
+1. [Contact list](3-apps/1-contact-list/README.md) - Example of a full stack application build with React, MongoDB and Jitar.
 1. [Microservices](3-apps/2-microservices/README.md) - Demonstrates how to setup a microservice architecture using Jitar.
